dermo-scope-next/pyproject.toml
[project]
name = "dermo-scope-lite"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "ai-edge-litert==2.1.4",
  "numpy>=1.26.0,<3.0",
  "Pillow>=10.0.0",
]
